Name change from KartesianProfile to SpatialProfile to reflect included Cylinder sampling capabilities.
