--- examples/general-plugins/ProfileOutput/profiles.xml
@@ -1,136 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<mardyn version="20100525">
-  <refunits type="SI">
-    <length unit="nm">0.0529177</length>
-    <mass unit="u">100</mass>
-    <energy unit="eV">27.2126</energy>
-  </refunits>
-  <simulation type="MD">
-    <integrator type="Leapfrog">
-      <timestep unit="reduced">0.0005</timestep>
-    </integrator>
-    <run>
-      <currenttime>0</currenttime>
-      <production>
-        <steps>100</steps>
-      </production>
-    </run>
-    <ensemble type="NVT">
-      <temperature unit="reduced">0.86232</temperature>
-      <domain type="box">
-        <lx>72.7626</lx>
-        <ly>62.9432</ly>
-        <lz>72.7626</lz>
-      </domain>
-      <components>
-        <moleculetype id="1" name="Test">
-          <site type="LJ126" id="1" >
-            <coords> <x>0.0</x> <y>0.0</y> <z>0.0</z> </coords>
-            <mass>1.0</mass>
-            <sigma>1.0</sigma>
-            <epsilon>1.0</epsilon>
-            <shifted>1</shifted>
-          </site>
-          <momentsofinertia rotaxes="xyz" >
-          <Ixx>0.0</Ixx>
-          <Iyy>0.0</Iyy>
-          <Izz>0.0</Izz>
-          </momentsofinertia>
-        </moleculetype>
-      </components>
-      <phasespacepoint>
-        <file type="ASCII">Ads_LJ1043.inp</file>
-      </phasespacepoint>
-    </ensemble>
-    <algorithm>
-      <parallelisation type="DomainDecomposition"/>
-      <datastructure type="LinkedCells">
-        <cellsInCutoffRadius>1</cellsInCutoffRadius>
-      </datastructure>
-
-      <cutoffs type="CenterOfMass">
-        <radiusLJ unit="reduced">2.5</radiusLJ>
-      </cutoffs>
-
-      <electrostatic type="ReactionField">
-        <epsilon>1.0e+10</epsilon>
-      </electrostatic>
-      <thermostats>
-        <thermostat type="TemperatureControl">
-          <control>
-            <start>0</start>
-            <frequency>1</frequency>
-            <stop>1000000000</stop>
-          </control>
-          <regions>
-            <region>
-              <method>Andersen</method>
-              <coords>
-                <lcx>0.0</lcx>
-                <lcy>0.0</lcy>
-                <lcz>0.0</lcz>
-                <ucx>72.7626</ucx>
-                <ucy>62.9432</ucy>
-                <ucz>72.7626</ucz>
-              </coords>
-              <target>
-                <temperature>293.44</temperature>
-                <component>1</component>
-              </target>
-              <settings>
-                <nu>0.05</nu>
-              </settings>
-            </region>
-          </regions>
-        </thermostat>
-      </thermostats>
-
-    </algorithm>
-
-    <output>
-      <outputplugin name="VTKMoleculeWriter">
-        <outputprefix>vtkOutput</outputprefix>
-        <writefrequency>100</writefrequency>
-      </outputplugin>
-    </output>
-
-    <plugin name="WallPotential">
-      <potential>104</potential>
-      <density>1.07</density>
-      <sigma>1</sigma>
-      <epsilon>100</epsilon>
-      <yoff>4.75</yoff>
-      <ycut>51</ycut>
-      <delta>0.775</delta>
-      <component id="1">
-        <xi>0.0514</xi>
-        <eta>1</eta>
-      </component>
-    </plugin>
-
-    <plugin name="Mirror">
-      <yPos>55</yPos>
-      <forceConstant>10</forceConstant>
-      <direction>0</direction>
-    </plugin>
-
-    <plugin name="KartesianProfile">
-      <x>1</x>
-      <y>20</y>
-      <z>20</z>
-      <writefrequency>100</writefrequency>
-      <timesteps>
-        <init>1</init>
-        <recording>1</recording>
-      </timesteps>
-      <outputprefix>comparison</outputprefix>
-      <profiles>
-        <density>1</density>
-        <temperature>1</temperature>
-        <velocity>1</velocity>
-        <velocity3d>1</velocity3d>
-      </profiles>
-    </plugin>
-
-  </simulation>
-</mardyn>
